As soon as you do this, you can replace the random string with BattleManager::TYPE_NO_JEDI. Below that, use BattleManager::TYPE_ONLY_JEDI.
That will work the exact same way as before. In index.php, do the same thing: <?php echo BattleManager::TYPE_NORMAL. Copy that and replace it with TYPE_NO_JEDI and TYPE_ONLY_JEDI.
To prove it still works, refresh this page. Everything's still happy!

In a sense, nothing changed! But now, these magic strings have a single home: at the top of BattleManager. If we ever needed to change these strings, we can do it in just once place.