Static versus Non Static Here's the deal: constants are static, and so far, all of our properties and methods are non-static.
You see, whenever you add something to a class - like a property or a method - you can choose to attach it to an individual instance of the object or to the class itself. When you choose to attach something to a class, it's said to be 'static'.

Let's look at a real example. In AbstractShip, the properties id, name, weaponPower and strength are not static:

 123 lines  lib/Model/AbstractShip.php
... lines 1 - 2
abstract class AbstractShip
{
    private {dollar_sign}id;
    private {dollar_sign}name;
    private {dollar_sign}weaponPower = 0;
    private {dollar_sign}strength = 0;
... lines 12 - 121
}
That means that if you have two Ship objects, each has a different id, name, weaponPower and strength. If you change the name in one Ship it does not affect any other ship objects.

But, if we were to change these properties to static - which is something you can do - then suddenly the name property would be global to all ships, meaning two ship objects could not have different names. This would be the one name for all AbstractShip.

Remember - a class is like a blueprint for a ship, and an object is like a real, physical ship. Since each real ship has a different name, it makes sense to make the {dollar_sign}name property non-static. This attaches the name to each individual object.

But other times, it may make sense to attach a property to the blueprint itself, meaning to the class. For example, suppose that the very design of the ships guarantees that each should have a minimum strength of 100. Since that is a property of ships in general, we might add a new private static property called {dollar_sign}minimumStrength and use that to prevent individual ships from setting their specific  lower than this.