Class Constants are Static
So, with properties or methods, you can choose static or non-static. But constants, well, they're static by their very nature. And that makes sense: the TYPE constants in BattleManager are global to the BattleMangaer class in general - it wouldn't make sense for them to be different for different objects.

When you reference something statically, you always reference it by saying the class name, ::, and then whatever you're referencing.