The Special Self Keyword
Before we try an example, there's another special property of static things. Notice that we're inside BattleManager and we're referencing the BattleManager class. If you want to, you can change this to, self::TYPE_NO_JEDI.

In the same way that  refers to the current object, self refers to the class that we're inside of. So this didn't change our behavior: it's just a nice shortcut.

Now, let's see a real-life static method in action.