In index.php, the three battle types are hard coded right in the HTML:
 138 lines  index.php
... lines 1 - 29
<html>
... lines 31 - 55
  <body>
    <div class='container'>
... lines 58 - 92
      <div class='battle-box center-block border'>
        <div>
          <form method='POST' action='/battle.php'>
... lines 96 - 119
            <div class='text-center'>
              <label for='battle_type'>Battle Type</label>
              <select name='battle_type' id='battle_type' class='form-control drp-dwn-width center-block'>
                <option value='<?php echo BattleManager::TYPE_NORMAL ?>'>Normal</option>
                <option value='<?php echo BattleManager::TYPE_NO_JEDI ?>'>No Jedi Powers</option>
                <option value='<?php echo BattleManager::TYPE_ONLY_JEDI ?>'>Only Jedi Powers</option>
              </select>
            </div>
... lines 128 - 131
          </form>
        </div>
      </div>
    </div>
  </body>
</html>
So what happens if we decide to add a fourth battle type to BattleManager. No problem: add a new constant, then update the battle() method logic for whatever cool thing this new type does.

But surprise! If we forget to also add the new type to index.php, then nobody's going to be able to use it. Really, I'd prefer BattleManager to be completely in charge of the battle types so that it's the only file I need to touch when something changes.