At first, why namespaces exist might not be obvious, so hold onto that question. Let's see how they work first.
Above any class, you can - if you want to - add a namespace keyword followed by some string. Like, Battle or something more complicated like Battle\HiGuys\NiceNameSpace. A namespace is a string, and you can give it different parts by separating each with a backslash \ - that's the slash that feels a little wrong when you type it - it's usually an escape character.

To keep things simple, just set the namespace to Battle for now.

As soon as we did that, we actually changed the name of this class: it is no longer called BattleManager. In fact, you can see that PhpStorm now highlights our code with an 'Undefined class BattleManager' error. Thanks to the namespace, the class is now called Battle\BattleManager.

Refresh to prove it. Great!

So... that's really it! When you add a namespace above a class, the full class name becomes that namespace, a \, and then class name. Every place we reference this class name will now need to change - like inside of Container. We'll do that in a few minutes - we've got a few other things to do first.