There is one other thing that you need to know with namespace: it's the mystical use statement.
When you want to reference a class, it's perfectly valid to type out the entire long class name right where you need to use it. But in practice, you won't see this very often. Instead, people typically add a use statement at the top of the file that references the full class name: Battle\BattleManager.
As soon as you do, when you need to work with the class, you can once again write out only the short class name.

To be clear, the use statement does not change how namespaces work: it's just a shortcut. When PHP executes this file, it sees class BattleManager and says:

Huh, BattleManager? Let me check all of the use statements at the top of this file.
PHP then looks to see if any of the use statements end in the word BattleManager. If it finds one, it basically copies the long class name and pastes it below right before executing the file. What I just did manually is what PHP basically does at run-time.

So use statements are just this nice, extra feature. And technically, you could avoid using them and instead write-out full class names right where you need them.

Ok! We're going to do a lot more with namespaces. But first, we need to turn to a very related topic called autoloading.