How could we make this function smarter? How could we make it automatically find new classes and files as we add them to the system?
Well I have an idea. BattleManager lives in the Service directory. What if we changed its namespace to match that? Or to get crazier, what if we gave every class a namespace that matches its directory?

If we did that, the autoload function could use the namespace to locate any file. The class - Service\BattleManager would live at Service/BattleManager.php. It's brilliant!

Now that we've changed the namespace to Service, we need to update any references to BattleManager - like in index.php. Change the use statement to Service.

Finally, in bootstrap.php, instead of manually checking for just this one class, say that the path is always equal to __DIR__/lib/ then str_replace() - we'll replace the back slash with a forward slash.

Notice I put two back-slashes. Since this is the escape character, if you only have one, it looks like you're escaping the next quote character. So, to get one slash, we need to use two - it's an ugly detail. Anyways, replace backslashes with a forward slash and pass that the class name. Finally, add the .php at the end.

Just in case, check to see if that file exists. If it does, require {dollar_sign}path.

That's it. Go back, refresh, and... everything still works.

And now, we are incredibly dangerous. We can now get rid of every single require statement really easily. Let's do it!