We could say Service\Container here or we can add a use statement for Service\Container. Let's do tha last. We'll have the same problem down below with BrokenShip: PhpStorm is trying to warn me! Add a use Model\BrokenShip to take care of that.
We'll probably have the same problem in battle.php - so open that up. Yep, add use Service\Container.

Looking good!

Reading the Error Messages... Closely

Try it again! Ok:

Class Service\RebelShip not found in ShipLoader.

Remember what I just said about reading the error messages closely? This one has a clue: it's looking for Service\RebelShip. But we don't have a class called Service\RebelShip - our class is called Model\RebelShip:

 38 lines  lib/Model/RebelShip.php
... lines 1 - 2
namespace Model;
class RebelShip extends AbstractShip
... lines 6 - 38

The problem exists where we're referencing this class - so in ShipLoader at line 43.

This is the most common mistake with namespaces: we have new RebelShip, but we don't had a use statement on top for this:

63 lines  lib/Service/ShipLoader.php
... lines 1 - 2
namespace Service;
class ShipLoader
{
... lines 7 - 40
    private function createShipFromData(array {dollar_sign}shipData)
    {
        if ({dollar_sign}shipData['team'] == 'rebel') {
            {dollar_sign}ship = new RebelShip({dollar_sign}shipData['name']);
        } else {
            {dollar_sign}ship = new Ship({dollar_sign}shipData['name']);
... line 47
        }
... lines 49 - 54
    }
... lines 56 - 60
}

This is the same problem we just solved in index.php, but with a small difference. Unlike index.php and battle.php, this file lives in a namespace called Service. That causes PHP to assume that RebelShip also lives in that namespace -- you know like roommates.

Here's how it works: when PHP parses this file, it sees the RebelShip class on line 43. Next, it looks up at the top of the file to see if there are any use statements that end in RebelShip. Since there aren't, it assumes that RebelShip also lives in the Service namespace, so Service\RebelShip.

Think about it: this is just like directories on your filesystem. If you are inside of a directory called Service and you say ls RebelShip, it's going to look for RebelShip inside of the Service directory.

But in index.php - since this doesn't hold a class - we didn't give this file a namespace. If you forget a use statement for BrokenShip here, this is equivalent to saying ls BrokenShip from the root of your file system, instead of from inside some directory.

In both cases the solution is the same: add the missing use statement: use Model\RebelShip.
We have the same problem below for Ship: add use Model\Ship.