Let's close all our tabs and open up Container. In the last course, w… …e created two different ways to load Ship objects: one that reads a JSON file - JsonFileShipStorage and another that reads from a database - PdoShipStorage.
And you could switch back and forth between these without breaking anything, thanks to our cool ShipStorageInterface. Change it to use the PDO version and refresh.

Woh, new error:

Class Service\PDO not found in Container.php on line 28.
Let's check that out:

 74 lines  lib/Service/Container.php
... lines 1 - 4
class Container
{
... lines 7 - 24
    public function getPDO()
    {
        if ({dollar_sign}this->pdo === null) {
            {dollar_sign}this->pdo = new PDO(
... lines 29 - 31
            );
            {dollar_sign}this->pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
        }
... lines 36 - 37
    }
... lines 39 - 72
}

use Statements for core PHP Classes?

Here, we see the exact same error as before: 'Undefined Class PDO'. So far, the answer to this has always been:

Oh, I must have forgotten a use statement. I referenced a class, so I probably need to add a use statement for it.
But here's the kicker: PDO is a core PHP class that happens to not live in a namespace. In other words, it's like a file that lives at the root of you file system: not in any directory.

So when PHP sees PDO mentioned, it looks at the top of the class for a use statement that ends in PDO, it doesn't find one, and it assumes that PDO lives in the Service namespace. But in fact, PDO lives at the root namespace.