Configuring Autoloading
To tell Composer to do the autoloading for us, all you need is a small configuration file called composer.json. Inside, add an autoload key, then a psr-4 key, and empty quotes set to lib.

That's it.

Remember how I said this rule is called PSR-0? Well PSR-4 is a slight amendment to PSR-0, but they both refer to the same thing. This tells Composer that we want to autoload using the PSR-0 convention, and that it should look for all classes inside the lib directory. That's it.