Let's talk about something totally different: a powerful part of object-oriented code called exceptions.
In index.php, we create a BrokenShip object. I'm going to do something crazy, guys. I'm going to say, {dollar_sign}brokenShip->setStrength() and pass it... banana.

That strength makes no sense. And if we try to battle using this ship, we should get some sort of error. But when we refresh... well, it is an error: but not exactly what I expected.

This error is coming from AbstractShip line 65. Open that up. I want you to look at 2 exceptional things here:

 125 lines  lib/Model/AbstractShip.php
... lines 1 - 4
abstract class AbstractShip
{
... lines 7 - 44
    public function setStrength({dollar_sign}number)
    {
      throw new Exception('Strength must be a number, duh!');
    }
      {dollar_sign}this->strength = {dollar_sign}number;
    }
... lines 53 - 123
}
First, we planned ahead. When we created the setStrength() method, we said:

You know what? This needs to be a number, so if somebody passes something dumb like 'banana,' then let's check for that and trigger an error.
Second, in order to trigger an error, we threw an exception. And that's actually what I want to talk about: Exceptions are classes, but they're completely special.