Second, exceptions are catchable. Here's what that means. Suppose that I wanted to kill the page right here with an error. I actually have two options: I can throw an exception, or I could print some error message and use a die statement to stop execution.
But when you use a die statement, your script is truly done: none of your other code executes. But with an exception, you can actually try to recover and keep going!

Let's look at how. Open up PdoShipStorage. Inside fetchAllShipsData(), change the table name to foooood.
That clearly will not work. This method is called by ShipLoader, inside getShips().

When we try to run this, we get an exception:

Base table or view not found
The error is coming from PdoShipStorage on line 11, but we can also see the line that called this: ShipLoader line 51.

Now, what if we knew that sometimes, for some reason, an exception like this might be thrown when we call fetchAllShipsData(). And when that happens, we don't want to kill the page or show an error. Instead, we want to - temporarily - render the page with zero ships.

How can we do this? First, surround the line - or lines - that might fail with a try-catch block. In the catch, add \Exception {dollar_sign}e. Now, if the fetchAllShipsData() method throws an exception, the page will not die. Instead, the code inside catch will be called and then execution will keep going like normal. That means, we can say {dollar_sign}shipData = array()