Using the Exception Object
And just like that, the page works. That's the power of exceptions. When you throw an exception, any code that calls your code has the opportunity to catch the exception and say:

No no no, I don't want the page to die. Instead, let's do something else.
Of course, we probably also don't want this to fail silently without us knowing, so you might trigger an error and print the message for our logs. Notice, in catch, we have access to the Exception object, and every exception has a getMessage() method on it. Use that to trigger an error to our logs.

Ok, refresh! Right now, we see the error on top of the page. But that's just because of our error_reporting settings in php.ini. On production, this wouldn't display, but would write a line to our logs.