Catching an exception is really powerful. But you can get even fancier.
For right now, var_dump() the Exception object.

Ok, this object is actually a PDOException. Two Important things: all exceptions are objects, and all exception classes ultimately extend PHP's base Exception class. So if you could look at the source code for PDOException, you'd see that it extends Exception:

And this ends up giving us a lot more flexibility when working with exceptions. Why? Remember, we're pretending that - for some reason - we occasionally have some database problems that cause a PDOException to be thrown. When that happens, we want to recover and just show zero ships. And we've got that.

But what if something else goes wrong inside fetchAllShipsData() that has nothing to do with talking to the database. Well, that would be truly unexpected, and in those cases, I want to let the exception be thrown like normal so we can see it while we're developing.