If I give you an object, could you print it? What I mean is, in battle.php, after we determine the winners, we echo {dollar_sign}ship1->getName(), which is of course a string. But could we just print {dollar_sign}ship1 and {dollar_sign}ship2?
Does it make sense to print an object? The answer is... no. Try to battle, you get a very clear error that says:

Object of class Model\RebelShip could not be converted to string in battle.php.
Remember this error: you'll eventually try to print an object on accident and see this!