Why am I telling you this seemingly small and obvious fact? Because I'm lying! You can print objects! You just have to do a little bit more work.
Here's the big picture: there are ways to give a class super-powers - like the ability to be printed or - as we'll see next - the ability to pretend like it's an array.

Open up AbstractShip. To make objects of this class printable, go to the bottom and create a new public function __toString().

Go back, refresh, and now it works just fine.

By adding the __toString() method - we gave PHP the ability to convert our object into a string. The __toString() must be called exactly like this, and there are other methods that take on special meaning. They all start with __, and we've already seen one: __construct():

 130 lines  lib/Model/AbstractShip.php
... lines 1 - 4
abstract class AbstractShip
{
... lines 7 - 29
  public function __construct({dollar_sign}name)
  {
    {dollar_sign}this->name = {dollar_sign}name;
  }
... lines 34 - 128
}
These are collectively called Magic Methods.