There are actually just a few magic methods: let's look at another common one. In battle.php, scroll down a little bit to where it shows the ship health. This should not work, and PHPStorm tells us why: the member - meaning property - has private access. We can't access a private property from outside the class.
But once again - via a magic method - you can bend the rules. This time, add a public function __get() with a single argument: . For now, just dump that.

Refresh to see what happens. Interesting! It dumps the string strength. Here's the magic: if you reference a property on your object that is not accessible - either because it doesn't exist or is private or protected - and you have an __get() method, then PHP will call that and pass you the property name.