After the last chapter, you might expect me to go into BattleResults and add some new magic method down at the bottom that would make this legal. But nope!
There is actually a second way to add special behavior to a class, and this method involves interfaces. Basically, PHP has a group of built-in interfaces and each gives your class a different super-power if you implement it.

The most famous is probably \ArrayAccess.

Of course as soon as you implement any interface, it will require you to add some methods. In this case, PhpStorm is telling me that I needed offsetGet(), offsetUnset(), offsetExist() and offsetSet().

Ok, let's do that, but with a little help from my editor. In PhpStorm, I can go to the 'Code'->'Generate' menu and select 'Implement Methods'. Select these 4.

Cool!

And just by doing this, it's legal to treat our object like an array. And when someone tries to access some array key - like winningShip - we'll just return that property instead.