So, for offsetExists(), use a function called property_exists() and pass it {dollar_sign}this and {dollar_sign}offset: that will be whatever key the user is trying to access.
For offsetGet(), return {dollar_sign}this->{dollar_sign}offset and in offsetSet(), say {dollar_sign}this->{dollar_sign}offset = {dollar_sign}value.

And finally - even though it would be weird from someone to unset one of our keys, let's make that legal by removing the property: unset({dollar_sign}this->{dollar_sign}offset).

Ok, this is a little weird, but it works. Now, just like with magic methods, don't run and use this everywhere for no reason. But occasionally, it might come in handy. And more importantly, you will see this sometimes in outside libraries. This means that even though something looks like an array, it might actually be an object.