Create a new PHP class called ShipCollection.
Hey, check it out: PhpStorm already correctly-guessed that this should have the Model namespace: it understands our PSR-0 naming convention.

Inside, add a private  property: this will be an array of Ship objects. Then add a public function __construct() method, give it a  argument, and set that property inside.

Above the  just to help our editor with autocompletion later, add some PHP Doc that says that this is an array of AbstractShip.