Obviously, ShipCollection is a class... but its only purpose is to be a small wrapper around an array. In ShipLoader, instead of returning the array, return a new ShipCollection() object and pass it {dollar_sign}ships.
Now, stop: we're referencing ShipCollection inside of ShipLoader, so we need a use statement for it. Go to the top to add it. But wait! It's already there! Thank you PhpStorm: it added it automatically for me when I auto-completed the class name. Whether your editor does this or not, just make sure to not forget those use statements!

Finally, above the method, we're not returning an array of AbstractShip objects anymore: we're now returning a ShipCollection.

Cool Now again, don't worry about why we're doing this yet. For now, let's try to fix our app.

First, go to index.php. Boom!

Cannot use object of type ShipCollection as array in index.php on line 13.
No surprise. After creating the , we're trying to add it to the ShipCollection as if it were an array!

 145 lines  index.php
... lines 1 - 11
{dollar_sign}ships = {dollar_sign}shipLoader->getShips();
{dollar_sign}brokenShip = new BrokenShip('Just a hunk of metal');
{dollar_sign}ships[] = {dollar_sign}brokenShip;