Ok, let's finally answer the question: why did we do this? Because sometimes, it might be useful to add some helpful methods to an array. Well, of course you can't do that, but you can add methods to a class.
For example, add a new method called public function removeAllBrokenShips(), because maybe we want a collection of only working ships. By adding this method, that would be really easy.

Let's test this fancy new method out. In index.php, call ->removeAllBrokenShips().

This looks and acts like an array, but with the super-power to have methods on it. ooOOOooo.

Refresh and check this out: no more broken ships, ever.

There are more of these interfaces that have special powers, but these are the most common ones. And the most important thing is just to understand that they exist and how they work.