Code Duplication
So, what's the problem? Look at BountyHunterShip and also look at Ship: there's some duplication. Both classes have a jediFactor property, a getJediFactor() method that returns this, and a setJediFactor that changes it.

Duplication is a bummer. How can we fix this? Well, we could use inheritance. But in this case, it's weird.

For example, we could make BountyHunterShip extend Ship, but then it would inherit this extra stuff that we don't really want or need. We could make it work, but I just don't like it.

Ok, what about making Ship extend BountyHunterShip? That just completely feels wrong: philosophically, not all Ships are BountyHunterShips - it's just not the right way to model these classes.

Are we stuck? What we want is a way to just share these 3 things: the jediFactor property, getJediFactor() and setJediFactor(). When you only need to share a few things, the right answer might be a trait.