Let's see what this trait thing is. In the Model directory, create a new PHP class called SettableJediFactorTrait. Now, change the class keyword to trait. Traits look and feel exactly like a normal class.
In fact, open up BountyHunterShip and move the property and first method into the trait. Also grab setJediFactor() and put that in the trait too.

The only difference between classes and traits is that traits can't be instantiated directly. Their purpose is for sharing code.

In BountyHunterShip, we can effectively copy and paste the contents of that trait into this class by going inside the class and adding use SettableJediFactorTrait.

That use statement has nothing to do with the namespace use statements: it's just a coincidence. As soon as we do this, when PHP runs, it will copy the contents of the trait and pastes them into this class right before it executes our code. It's as if all the code from the trait actually lives inside this class.