In modern PHP, you're going to spend a lot of time working with other people's classes: via external libraries that you bring into your project to get things done faster. Of course, when you do that: you can't actually edit their code if you need to change or add some behavior.
Fortunately, OO code gives us some really neat ways to deal with this limitation.

Modifying a Class without Modifying it?

For the next few minutes, I want you to pretend like our PDOShipStorage is actually from a third-party library. In other words, we can't modify it.

Now, let's say whenever we call fetchAllShipsData(), it's really important for us to log to a file, how many ships were found. But if we can't edit this file, how can we do that?

Using Inheritance

There's actually two ways to do this, and both are pretty awesome. The first way is to create a new class that extends PDOShipStorage, like LoggablePDOShipStorage, and override some methods to add logging.