To use the new class, open up Container. Inside getShipStorage(), add {dollar_sign}this->shipStorage = new LoggableShipStorage() and pass it {dollar_sign}this->shipStorage, which is the PDOShipStorage object.
We've just pulled a 'fast one' on our application: our entire app thinks we're using PDOShipStorage, but we just changed that! If you refresh now, nothing is different: everything still eventually goes through the PDOShipStorage object.

But now, we have the opportunity to add more functionality - or to change functionality - in either of these methods.